uniforum-backend/password-service/requirements.txt
fastapi
uvicorn
pika
python-dotenv
email-validator